refactor(codegen): get runtimeApis from metadata, tweaks name, optimize code
