+++ packages/api/src/index.ts
@@ -1,5 +1,6 @@
 import DelightfulApi from './DelightfulApi';
 
 export type * from './types';
+export * from './executor';
 export { DelightfulApi };
 export { HttpProvider, ScProvider, WsProvider } from '@polkadot/rpc-provider';
