+++ packages/chaintypes/src/moonbeam/index.ts
@@ -6,7 +6,7 @@ import { ChainStorage } from './query';
 import { RpcCalls } from './rpc';
 import { ChainErrors } from './errors';
 import { ChainEvents } from './events';
-import { RuntimeCalls } from './call';
+import { RuntimeCalls } from './runtime';
 
 export * from './types';
 export * from './consts';
