--- packages/chaintypes/src/substrate/call.ts
@@ -1,102 +1,0 @@
-// Generated by @delightfuldot/codegen
-
-import type { GenericRuntimeCalls, GenericRuntimeCall } from '@delightfuldot/types';
-import type {
-  BabeConfiguration,
-  Epoch,
-  Slot,
-  Option,
-  OpaqueKeyOwnershipProof,
-  AccountId32Like,
-  AccountId32,
-} from '@delightfuldot/codecs';
-
-export interface RuntimeCalls extends GenericRuntimeCalls {
-  /**
-   * @module: babe
-   * @runtimehash: 0xcbca25e39f142387
-   * @version: 2
-   **/
-  BabeApi: {
-    /**
-     * Return the configuration for BABE.
-     *
-     * @call: BabeApi_configuration
-     **/
-    configuration: GenericRuntimeCall<() => Promise<BabeConfiguration>>;
-
-    /**
-     * Returns information regarding the current epoch.
-     *
-     * @call: BabeApi_current_epoch
-     **/
-    currentEpoch: GenericRuntimeCall<() => Promise<Epoch>>;
-
-    /**
-     * Returns the slot that started the current epoch.
-     *
-     * @call: BabeApi_current_epoch_start
-     **/
-    currentEpochStart: GenericRuntimeCall<() => Promise<Slot>>;
-
-    /**
-     * Returns information regarding the next epoch (which was already previously announced).
-     *
-     * @call: BabeApi_next_epoch
-     **/
-    nextEpoch: GenericRuntimeCall<() => Promise<Epoch>>;
-
-    /**
-     * Generates a proof of key ownership for the given authority in the
-     * current epoch. An example usage of this module is coupled with the
-     * session historical module to prove that a given authority key is
-     * tied to a given staking identity during a specific session. Proofs
-     * of key ownership are necessary for submitting equivocation reports.
-     * NOTE: even though the API takes a `slot` as parameter the current
-     * implementations ignores this parameter and instead relies on this
-     * method being called at the correct block height, i.e. any point at
-     * which the epoch for the given slot is live on-chain. Future
-     * implementations will instead use indexed data through an offchain
-     * worker, not requiring older states to be available.
-     *
-     * @call: BabeApi_generate_key_ownership_proof
-     **/
-    generateKeyOwnershipProof: GenericRuntimeCall<
-      (slot: Slot, authorityId: AccountId32Like) => Promise<Option<OpaqueKeyOwnershipProof>>
-    >;
-
-    /**
-     * Submits an unsigned extrinsic to report an equivocation. The caller
-     * must provide the equivocation proof and a key ownership proof
-     * (should be obtained using `generate_key_ownership_proof`). The
-     * extrinsic will be unsigned and should only be accepted for local
-     * authorship (not to be broadcast to the network). This method returns
-     * `None` when creation of the extrinsic fails, e.g. if equivocation
-     * reporting is disabled for the given runtime (i.e. this method is
-     * hardcoded to return `None`). Only useful in an offchain context.
-     *
-     * @call: BabeApi_submit_report_equivocation_unsigned_extrinsic
-     **/
-    submitReportEquivocationUnsignedExtrinsic: GenericRuntimeCall<
-      (equivocationProof: BabeConfiguration, keyOwnerProof: OpaqueKeyOwnershipProof) => Promise<Option<null>>
-    >;
-
-    [method: string]: GenericRuntimeCall;
-  };
-
-  /**
-   * @module: discovery
-   * @runtimehash: 0x687ad44ad37f03c2
-   * @version: 1
-   **/
-  AuthorityDiscoveryApi: {
-    /**
-     * Retrieve authority identifiers of the current and next authority set.
-     *
-     * @call: AuthorityDiscoveryApi_authorities
-     **/
-    authorities: GenericRuntimeCall<() => Promise<Array<AccountId32>>>;
-
-    [method: string]: GenericRuntimeCall;
-  };
-}
