+++ packages/codegen/src/index.ts
@@ -2,7 +2,7 @@ import { DelightfulApi } from 'delightfuldot';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as process from 'process';
-import { ConstsGen, ErrorsGen, EventsGen, IndexGen, QueryGen, RpcGen, TypesGen, CallGen } from './generator';
+import { ConstsGen, ErrorsGen, EventsGen, IndexGen, QueryGen, RpcGen, TypesGen, RuntimeCallsGen } from './generator';
 import { RpcMethods } from '@delightfuldot/types';
 import { MetadataLatest } from '@delightfuldot/codecs';
 import { NetworkInfo } from './types';
@@ -10,8 +10,9 @@ import { NetworkInfo } from './types';
 export async function generateTypesFromChain(network: NetworkInfo, endpoint: string, outDir: string) {
   const api = await DelightfulApi.create(endpoint);
   const { methods }: RpcMethods = await api.rpc.rpc.methods();
+  const { apis } = api.runtimeVersion;
 
-  await generateTypes(network, api.metadataLatest, methods, outDir, api.runtimeVersion.apis);
+  await generateTypes(network, api.metadataLatest, methods, apis, outDir);
 
   await api.disconnect();
 }
@@ -20,8 +21,8 @@ export async function generateTypes(
   network: NetworkInfo,
   metadata: MetadataLatest,
   rpcMethods: string[],
+  runtimeApis: [string, number][],
   outDir: string = '.',
-  apis: [string, number][],
 ) {
   const dirPath = path.resolve(process.cwd(), outDir, network.chain);
   const defTypesFileName = path.join(dirPath, `types.ts`);
@@ -31,7 +32,7 @@ export async function generateTypes(
   const indexFileName = path.join(dirPath, `index.ts`);
   const errorsFileName = path.join(dirPath, `errors.ts`);
   const eventsFileName = path.join(dirPath, `events.ts`);
-  const runtimeCallsFileName = path.join(dirPath, `call.ts`);
+  const runtimeCallsFileName = path.join(dirPath, `runtime.ts`);
 
   if (!fs.existsSync(dirPath)) {
     fs.mkdirSync(dirPath);
@@ -44,7 +45,7 @@ export async function generateTypes(
   const indexGen = new IndexGen(network);
   const errorsGen = new ErrorsGen(typesGen);
   const eventsGen = new EventsGen(typesGen);
-  const callGen = new CallGen(typesGen, apis);
+  const runtimeCallsGen = new RuntimeCallsGen(typesGen, runtimeApis);
 
   fs.writeFileSync(defTypesFileName, await typesGen.generate());
   fs.writeFileSync(errorsFileName, await errorsGen.generate());
@@ -53,5 +54,5 @@ export async function generateTypes(
   fs.writeFileSync(queryTypesFileName, await queryGen.generate());
   fs.writeFileSync(constsTypesFileName, await constsGen.generate());
   fs.writeFileSync(indexFileName, await indexGen.generate());
-  fs.writeFileSync(runtimeCallsFileName, await callGen.generate());
+  fs.writeFileSync(runtimeCallsFileName, await runtimeCallsGen.generate());
 }
